refactor(codegen): extrai passes/parallelism.rs de func.rs
Move silent parallelism Level-1 (~1432 linhas, 11 fns) para
crates/rts-codegen/src/codegen/lower/passes/parallelism.rs:
- lift_inline_arrows_in_array_methods + try_lift_arrow_arg + helpers
- array_methods_pass + rewrite_array_methods_in_{stmt,expr}
- reduce_pass + apply_reduce_pass_to_{top_level,body} + AssocOp
  + make_par_reduce_expr
- purity_pass + apply_purity_pass_to_{top_level,body} + helpers
  (build_pure_ns_set, is_pure_expr_for_parallel, analyze_for_of_body_pure,
  make_par_foreach_stmt, is_known_global_ident, has_capture)

func.rs: 5194 -> 3763 linhas. Suite verde (11/11 crates).
